Move the next batch of structs into headers
