+++ src/OpenLoco/src/Game.cpp
@@ -2,6 +2,7 @@
 #include "Audio/Audio.h"
 #include "Config.h"
 #include "GameCommands/GameCommands.h"
+#include "GameCommands/General/LoadSaveQuit.h"
 #include "GameException.hpp"
 #include "GameState.h"
 #include "GameStateFlags.h"
