+++ src/OpenLoco/src/GameCommands/General/LoadSaveQuit.cpp
@@ -1,3 +1,4 @@
+#include "LoadSaveQuit.h"
 #include "Audio/Audio.h"
 #include "Game.h"
 #include "GameCommands/GameCommands.h"
