+++ src/OpenLoco/src/GameCommands/General/RenameStation.cpp
@@ -1,3 +1,4 @@
+#include "RenameStation.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
 #include "Graphics/Gfx.h"
