+++ src/OpenLoco/src/GameCommands/General/SetGameSpeed.cpp
@@ -1,3 +1,4 @@
+#include "SetGameSpeed.h"
 #include "GameCommands/GameCommands.h"
 #include "SceneManager.h"
 #include <OpenLoco/Interop/Interop.hpp>
