+++ src/OpenLoco/src/GameCommands/Industries/RemoveIndustry.cpp
@@ -1,3 +1,4 @@
+#include "RemoveIndustry.h"
 #include "Economy/Economy.h"
 #include "Economy/Expenditures.h"
 #include "GameCommands/GameCommands.h"
