+++ src/OpenLoco/src/GameCommands/Industries/RemoveIndustry.h
@@ -2,5 +2,25 @@
 
 namespace OpenLoco::GameCommands
 {
+    struct IndustryRemovalArgs
+    {
+        static constexpr auto command = GameCommand::removeIndustry;
+
+        IndustryRemovalArgs() = default;
+        explicit IndustryRemovalArgs(const registers& regs)
+            : industryId(static_cast<IndustryId>(regs.dl))
+        {
+        }
+
+        IndustryId industryId;
+
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.dl = enumValue(industryId);
+            return regs;
+        }
+    };
+
     void removeIndustry(registers& regs);
 }
