+++ src/OpenLoco/src/Windows/TimePanel.cpp
@@ -1,6 +1,8 @@
 #include "Date.h"
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "GameCommands/GameCommands.h"
+#include "GameCommands/General/SetGameSpeed.h"
+#include "GameCommands/General/TogglePause.h"
 #include "Graphics/Colour.h"
 #include "Graphics/Gfx.h"
 #include "Graphics/ImageIds.h"
