+++ src/OpenLoco/src/Windows/TitleMenu.cpp
@@ -3,6 +3,7 @@
 #include "EditorController.h"
 #include "Game.h"
 #include "GameCommands/GameCommands.h"
+#include "GameCommands/General/LoadSaveQuit.h"
 #include "Graphics/Colour.h"
 #include "Graphics/Gfx.h"
 #include "Graphics/ImageIds.h"
