+++ src/OpenLoco/src/Windows/ToolbarTop.cpp
@@ -4,6 +4,7 @@
 #include "Entities/EntityManager.h"
 #include "Game.h"
 #include "GameCommands/GameCommands.h"
+#include "GameCommands/General/LoadSaveQuit.h"
 #include "Graphics/Colour.h"
 #include "Graphics/Gfx.h"
 #include "Graphics/ImageIds.h"
